feat: Seperate file-upload part to provider
