+++ src/domain/file/providers/file.repository.ts
@@ -1,9 +1,10 @@
-import { UploadFile } from '#entities/file';
 import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+// import { UploadFile } from '#entities/file';
+
+// import { InjectRepository } from '@nestjs/typeorm';
+// import { Repository } from 'typeorm';
 
 @Injectable()
 export class FilesRepository {
-  constructor(@InjectRepository(UploadFile) private filesRepository: Repository<UploadFile>) {}
+  //constructor(@InjectRepository(UploadFile) private filesRepository: Repository<UploadFile>) {}
 }
