+++ src/domain/file/providers/index.ts
@@ -1,0 +1,2 @@
+export * from './file.service';
+export * from './file.repository';
